requirements.txt
openai
python-dotenv
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
statsmodels

